requirements.txt
requests
python-dotenv
psycopg2  # or psycopg2-binary if needed for PostgreSQL connection